Potentially fixing activeClassName error
I get this error while using these components:

index.js:2178 Warning: React does not recognize the `activeClassName` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `activeclassname` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
    in div (at App.jsx:22)
    in div (created by RoutedTabs)
    in RoutedTabs (at App.jsx:17)
    in div (at App.jsx:15)
    in App (at index.js:12)
    in Router (created by BrowserRouter)
    in BrowserRouter (at index.js:11)

Looks like the fix was talked about on react router:
https://github.com/ReactTraining/react-router/issues/4908

What do you think?